+++ lib/config/sfFilterConfigHandler.class.php
@@ -33,7 +33,7 @@ class sfFilterConfigHandler extends sfYamlConfigHandler
   public function execute($configFiles)
   {
     // parse the yaml
-    $config = self::getConfiguration($configFiles);
+    $config = static::getConfiguration($configFiles);
 
     // init our data and includes arrays
     $data     = array();
@@ -169,14 +169,14 @@ EOF;
    */
   static public function getConfiguration(array $configFiles)
   {
-    $config = self::parseYaml($configFiles[0]);
+    $config = static::parseYaml($configFiles[0]);
     foreach (array_slice($configFiles, 1) as $i => $configFile)
     {
       // we get the order of the new file and merge with the previous configurations
       $previous = $config;
 
       $config = array();
-      foreach (self::parseYaml($configFile) as $key => $value)
+      foreach (static::parseYaml($configFile) as $key => $value)
       {
         $value = (array) $value;
         $config[$key] = isset($previous[$key]) ? sfToolkit::arrayDeepMerge($previous[$key], $value) : $value;
@@ -192,13 +192,13 @@ EOF;
       }
     }
 
-    $config = self::replaceConstants($config);
+    $config = static::replaceConstants($config);
 
     foreach ($config as $category => $keys)
     {
       if (isset($keys['file']))
       {
-        $config[$category]['file'] = self::replacePath($keys['file']);
+        $config[$category]['file'] = static::replacePath($keys['file']);
       }
     }